+++ src/scd41/scd41.h
@@ -17,5 +17,5 @@ int16_t scd41_reinit(void);
 int16_t scd41_get_serial_number(uint16_t *serial_0, uint16_t *serial_1, uint16_t *serial_2);
 int16_t scd41_start_periodic_measurement(void);
 int16_t scd41_stop_periodic_measurement(void);
-int16_t scd41_get_measures(uint16_t *co2, uint32_t *temperature, uint16_t *humidity);
+int16_t scd41_get_measures(uint16_t *co2, uint16_t *temperature, uint16_t *humidity);
 void scd41_wake_up(void);
